+++ tests/test_newconceptdesigns.py
@@ -30,22 +30,22 @@ class MITCalc(unittest.TestCase):
     def test_Closed(self):
         self.assertEqual(dimstack.utils.nround(stack.Closed.nominal), 9.8480)
         self.assertEqual(dimstack.utils.nround(stack.Closed.tolerance.T / 2), 0.0200)
-        self.assertEqual(dimstack.utils.nround(stack.Closed.Z_min), 9.8280)
-        self.assertEqual(dimstack.utils.nround(stack.Closed.Z_max), 9.8680)
+        self.assertEqual(dimstack.utils.nround(stack.Closed.abs_lower), 9.8280)
+        self.assertEqual(dimstack.utils.nround(stack.Closed.abs_upper), 9.8680)
 
     def test_WC(self):
         self.assertEqual(dimstack.utils.nround(stack.WC.nominal), 9.8480)
         self.assertEqual(dimstack.utils.nround(stack.WC.tolerance.T / 2), 0.0200)
-        self.assertEqual(dimstack.utils.nround(stack.WC.Z_min), 9.8280)
-        self.assertEqual(dimstack.utils.nround(stack.WC.Z_max), 9.8680)
+        self.assertEqual(dimstack.utils.nround(stack.WC.abs_lower), 9.8280)
+        self.assertEqual(dimstack.utils.nround(stack.WC.abs_upper), 9.8680)
 
     def test_SixSigma(self):
         self.assertEqual(dimstack.utils.nround(stack.SixSigma(at=6).nominal), 9.8480)
         self.assertEqual(dimstack.utils.nround(stack.SixSigma(at=6).nominal), 9.8480)
         self.assertEqual(dimstack.utils.nround(stack.SixSigma(at=6).tolerance.T / 2, 4), 0.0122)
         self.assertEqual(dimstack.utils.nround(stack.SixSigma(at=6).distribution.stdev * 2, 5), 0.00408)  # times 2 !?!?
-        self.assertEqual(dimstack.utils.nround(stack.SixSigma(at=6).Z_min, 4), 9.8358)
-        self.assertEqual(dimstack.utils.nround(stack.SixSigma(at=6).Z_max, 4), 9.8602)
+        self.assertEqual(dimstack.utils.nround(stack.SixSigma(at=6).abs_lower, 4), 9.8358)
+        self.assertEqual(dimstack.utils.nround(stack.SixSigma(at=6).abs_upper, 4), 9.8602)
 
     # def test_SixSigma_assembly(self):
     #     dim = stack.SixSigma(at=4.5)