+++ tests/test_tolerance_types.py
@@ -23,5 +23,52 @@ class FlippedBilateral(unittest.TestCase):
         self.assertEqual(t.lower, -0.005)
 
 
+class AbsRelPosNeg(unittest.TestCase):
+    def test_Positive_Abs(self):
+        d = dimstack.dim.Basic(
+            nom=0.1,
+            tol=dimstack.tolerance.SymmetricBilateral(0.2),
+            name="B",
+            desc="Washer Length",
+        )
+        self.assertAlmostEqual(d.abs_upper, 0.3)
+        self.assertEqual(d.abs_lower, -0.1)
+
+        self.assertAlmostEqual(d.rel_upper, 0.3)
+        self.assertEqual(d.rel_lower, -0.1)
+
+    def test_Negative_Abs(self):
+        d = dimstack.dim.Basic(
+            nom=-0.1,
+            tol=dimstack.tolerance.SymmetricBilateral(0.2),
+            name="B",
+            desc="Washer Length",
+        )
+        self.assertAlmostEqual(d.abs_upper, 0.1)
+        self.assertAlmostEqual(d.abs_lower, -0.3) # Edge Case
+
+        # self.assertEqual(d.rel_upper, 0.3)
+        # self.assertEqual(d.rel_lower, -0.1)
+
+    def test_Positive_Rel(self):
+        d = dimstack.dim.Basic(
+            nom=0.5,
+            tol=dimstack.tolerance.UnequalBilateral(upper=0.1, lower=-0.2),
+            name="B",
+            desc="Washer Length",
+        )
+        self.assertAlmostEqual(d.rel_upper, 0.6)
+        self.assertAlmostEqual(d.rel_lower, 0.3)
+
+    def test_Negative_Rel(self):
+        d = dimstack.dim.Basic(
+            nom=-0.5,
+            tol=dimstack.tolerance.UnequalBilateral(upper=0.1, lower=-0.2),
+            name="B",
+            desc="Washer Length",
+        )
+        self.assertEqual(d.rel_upper, 0.6)
+        self.assertAlmostEqual(d.rel_lower, 0.3)
+
 if __name__ == "__main__":
     unittest.main()